EWPP-1434: Move functionality from legacy oe_content_sub_entity_extra_authors to newly created oe_content_sub_entity_author.
