+++ modules/oe_content_person/tests/src/Kernel/PersonEntityTestBase.php
@@ -35,6 +35,7 @@ abstract class PersonEntityTestBase extends EntityKernelTestBase {
     'oe_content_departments_field',
     'oe_content_entity_contact',
     'oe_content_person',
+    'oe_content_sub_entity',
     'oe_content_social_media_links_field',
     'oe_content_sub_entity_document_reference',
     'oe_content_timeline_field',
