+++ modules/oe_content_publication/tests/src/Kernel/PublicationEntityTest.php
@@ -39,6 +39,7 @@ class PublicationEntityTest extends EntityKernelTestBase {
     'oe_content_entity_contact',
     'oe_content_publication',
     'oe_content_social_media_links_field',
+    'oe_content_sub_entity',
     'oe_content_sub_entity_document_reference',
     'oe_content_timeline_field',
     'file_link',
