+++ modules/oe_content_sub_entity/modules/oe_content_sub_entity_author/README.md
@@ -1,6 +1,27 @@
 # Author entity type
 
-At the moment, this module provides the following bundles:
-- "Corporate body" with possibility to reference some corporate body term.
+This module ships a set of various "Author" types which are used to introduce entities like Corporate body,Person, Organisation, or simple links to internal or external content as authors of the publications.
 
-When used on a [entity reference revision](https://www.drupal.org/project/entity_reference_revisions) field, in combination with the [inline entity form](https://www.drupal.org/project/inline_entity_form) widget, the "Author" entity type brings the additional value of being able to arrange heterogeneous author references (as in: references to different entity types and bundle) in the needed order.
+After enabling of module you should see additional "Author" types:
+- Corporate body (with reference some Corporate body term)
+- Person (with reference to some existing Person content)
+- Organisation (with reference to some existing Organisation content)
+- Link (with link field where you can add internal or external link)
+
+## Installation
+
+Before enabling this module, make sure the following modules are present in your codebase by adding them to your
+`composer.json` and by running `composer update`:
+
+```json
+"require": {
+  "drupal/composite_reference": "~1.0-alpha2",
+  "drupal/entity_reference_revisions": "~1.3",
+  "drupal/field_group": "~3.0",
+  "drupal/inline_entity_form": "~1.0-rc9",
+}
+```
+
+## Usage
+When used on a [entity reference revision](https://www.drupal.org/project/entity_reference_revisions) field, in combination with the [inline entity form](https://www.drupal.org/project/inline_entity_form) widget, the "Author" entity type brings the additional value of being able to arrange heterogeneous authors (as in: references to different entity types and bundle or external/internal links) in the needed order.
+To add support to these references in created "Authors" fields, you have to explicitly allow needed author types in the field configurations.
