+++ modules/oe_content_sub_entity/modules/oe_content_sub_entity_author/config/install/core.entity_form_display.oe_author.oe_organisation.default.yml
@@ -3,7 +3,7 @@ status: true
 dependencies:
   config:
     - field.field.oe_author.oe_organisation.oe_node_reference
-    - oe_content.oe_author_type.oe_organisation
+    - oe_content_sub_entity_author.oe_author_type.oe_organisation
 id: oe_author.oe_organisation.default
 targetEntityType: oe_author
 bundle: oe_organisation
