+++ modules/oe_content_sub_entity/modules/oe_content_sub_entity_author/config/install/core.entity_form_display.oe_author.oe_person.default.yml
@@ -3,7 +3,7 @@ status: true
 dependencies:
   config:
     - field.field.oe_author.oe_person.oe_node_reference
-    - oe_content.oe_author_type.oe_person
+    - oe_content_sub_entity_author.oe_author_type.oe_person
 id: oe_author.oe_person.default
 targetEntityType: oe_author
 bundle: oe_person
