+++ modules/oe_content_sub_entity/modules/oe_content_sub_entity_author/config/install/core.entity_view_display.oe_author.oe_link.default.yml
@@ -3,7 +3,7 @@ status: true
 dependencies:
   config:
     - field.field.oe_author.oe_link.oe_link
-    - oe_content.oe_author_type.oe_link
+    - oe_content_sub_entity_author.oe_author_type.oe_link
   module:
     - link
 id: oe_author.oe_link.default
@@ -13,11 +13,14 @@ mode: default
 content:
   oe_link:
     weight: 0
+    label: above
     settings:
-      placeholder_url: ''
-      placeholder_title: ''
+      trim_length: 80
+      url_only: false
+      url_plain: false
+      rel: ''
+      target: ''
     third_party_settings: {  }
-    type: link_default
+    type: link
     region: content
-hidden:
-  status: true
+hidden: {  }
