+++ modules/oe_content_sub_entity/modules/oe_content_sub_entity_author/config/install/oe_content_sub_entity_author.oe_author_type.oe_person.yml
