+++ modules/oe_content_sub_entity/modules/oe_content_sub_entity_author/tests/src/Kernel/AuthorEntityTest.php
@@ -5,80 +5,200 @@ declare(strict_types = 1);
 namespace Drupal\Tests\oe_content_sub_entity_author\Kernel;
 
 use Drupal\KernelTests\Core\Entity\EntityKernelTestBase;
-use Drupal\Tests\node\Traits\ContentTypeCreationTrait;
 use Drupal\Tests\node\Traits\NodeCreationTrait;
-use Drupal\Tests\field\Traits\EntityReferenceTestTrait;
 
 /**
- * Tests author reference entity.
+ * Tests all author entity bundles.
  */
 class AuthorEntityTest extends EntityKernelTestBase {
 
   use NodeCreationTrait;
-  use ContentTypeCreationTrait;
-  use EntityReferenceTestTrait;
 
   /**
    * {@inheritdoc}
    */
   public static $modules = [
     'node',
+    'link',
+    'entity_reference_revisions',
+    'sparql_entity_storage',
+    'rdf_skos',
+    'composite_reference',
+    'media',
+    'image',
+    'text',
+    'options',
+    'datetime',
+    'typed_link',
+    'field_group',
+    'maxlength',
+    'description_list_field',
+    'inline_entity_form',
+    'oe_content_entity_contact',
+    'oe_content_social_media_links_field',
+    'oe_content',
+    'oe_content_departments_field',
+    'oe_content_person',
+    'oe_content_timeline_field',
+    'oe_content_organisation',
+    'oe_content_sub_entity',
+    'oe_content_sub_entity_document_reference',
     'oe_content_sub_entity_author',
   ];
 
   /**
    * {@inheritdoc}
    */
-  protected function setUp(): void {
+  protected function setUp() {
     parent::setUp();
 
     $this->installEntitySchema('oe_author');
+    $this->installEntitySchema('oe_document_reference');
     $this->installEntitySchema('node');
-    $this->installConfig(['node', 'filter']);
-
-    // Create a node type.
-    $this->createContentType([
-      'type' => 'page',
-      'name' => 'Basic page',
+    $this->installEntitySchema('media');
+    $this->installConfig([
+      'node',
+      'filter',
+      'media',
+      'oe_content_social_media_links_field',
+      'oe_content',
+      'oe_content_departments_field',
+      'oe_content_person',
+      'oe_content_organisation',
+      'oe_content_sub_entity_author',
     ]);
-
-    // Create test bundle of Author reference entity.
-    $type_storage = $this->container->get('entity_type.manager')->getStorage('oe_author_type');
-    $type_storage->create([
-      'id' => 'test_bundle',
-      'label' => 'Test bundle',
-    ])->save();
-    $this->createEntityReferenceField('oe_author', 'test_bundle', 'field_reference_1', 'Node reference 1', 'node');
-    $this->createEntityReferenceField('oe_author', 'test_bundle', 'field_reference_2', 'Node reference 1', 'node');
-
   }
 
   /**
-   * Tests label of Author reference entity.
+   * Tests label of extra author entities.
    */
-  public function testLabel() {
-    // Author reference shows bundle label if there aren't referenced
-    // entities.
-    $author_reference = $this->entityTypeManager->getStorage('oe_author')->create([
-      'type' => 'test_bundle',
-    ]);
-    $author_reference->save();
-    $this->assertEquals('Test bundle', $author_reference->label());
+  public function testLabel(): void {
+    foreach ($this->getAuthorEntitiesTestData() as $bundle => $data_case) {
+      $author = $this->entityTypeManager->getStorage('oe_author')->create([
+        'type' => $bundle,
+      ]);
+      foreach ($data_case as $data) {
+        $field_values = NULL;
+        if (!empty($data['referenced_nodes'])) {
+          foreach ($data['referenced_nodes'] as $node_values) {
+            $field_values[] = $this->createNode($node_values);
+          }
+        }
+        elseif (!empty($data['field_values'])) {
+          $field_values = $data['field_values'];
+        }
+        $author->set($data['reference_field_name'], $field_values);
+        $author->save();
+        $this->assertEquals($data['expected_label'], $author->label());
+      }
+    }
+  }
 
-    // Author reference shows referenced entity label if it exists.
-    $node_1 = $this->createNode([
-      'title' => 'Referenced node 1 label',
-    ]);
-    $author_reference->set('field_reference_1', [$node_1])->save();
-    $this->assertEquals('Referenced node 1 label', $author_reference->label());
-
-    // Author reference shows referenced entity labels separated by comma,
-    // if they exist.
-    $node_2 = $this->createNode([
-      'title' => 'Referenced node 2 label',
-    ]);
-    $author_reference->set('field_reference_2', [$node_2])->save();
-    $this->assertEquals('Referenced node 1 label, Referenced node 2 label', $author_reference->label());
+  /**
+   * Testing data for author entity type.
+   *
+   * @return array
+   *   Array of test cases.
+   */
+  protected function getAuthorEntitiesTestData(): array {
+    return [
+      'oe_person' => [
+        [
+          'reference_field_name' => 'oe_node_reference',
+          'referenced_nodes' => NULL,
+          'expected_label' => 'Person',
+        ],
+        [
+          'reference_field_name' => 'oe_node_reference',
+          'referenced_nodes' => [
+            [
+              'type' => 'oe_person',
+              'oe_person_first_name' => 'John',
+              'oe_person_last_name' => 'Doe',
+            ],
+          ],
+          'expected_label' => 'John Doe',
+        ],
+        [
+          'reference_field_name' => 'oe_node_reference',
+          'referenced_nodes' => [
+            [
+              'type' => 'oe_person',
+              'oe_person_first_name' => 'John',
+              'oe_person_last_name' => 'Doe',
+            ],
+            [
+              'type' => 'oe_person',
+              'oe_person_first_name' => 'Foo',
+              'oe_person_last_name' => 'Bar',
+            ],
+          ],
+          'expected_label' => 'John Doe, Foo Bar',
+        ],
+      ],
+      'oe_organisation' => [
+        [
+          'reference_field_name' => 'oe_node_reference',
+          'referenced_nodes' => NULL,
+          'expected_label' => 'Organisation',
+        ],
+        [
+          'reference_field_name' => 'oe_node_reference',
+          'referenced_nodes' => [
+            [
+              'type' => 'oe_organisation',
+              'title' => 'European Commission',
+            ],
+          ],
+          'expected_label' => 'European Commission',
+        ],
+        [
+          'reference_field_name' => 'oe_node_reference',
+          'referenced_nodes' => [
+            [
+              'type' => 'oe_organisation',
+              'title' => 'Org 1',
+            ],
+            [
+              'type' => 'oe_organisation',
+              'title' => 'Org 2',
+            ],
+          ],
+          'expected_label' => 'Org 1, Org 2',
+        ],
+      ],
+      'oe_link' => [
+        [
+          'reference_field_name' => 'oe_link',
+          'field_values' => NULL,
+          'expected_label' => 'Link',
+        ],
+        [
+          'reference_field_name' => 'oe_link',
+          'field_values' => [
+            [
+              'title' => 'Link Title',
+              'uri' => 'http://example.com',
+            ],
+          ],
+          'expected_label' => 'Link Title',
+        ],
+        [
+          'reference_field_name' => 'oe_link',
+          'field_values' => [
+            [
+              'title' => 'Link Title 1',
+              'uri' => 'http://example.com',
+            ],
+            [
+              'title' => 'Link Title 2',
+              'uri' => 'http://example.com',
+            ],
+          ],
+          'expected_label' => 'Link Title 1, Link Title 2',
+        ],
+      ],
+    ];
   }
 
 }
