+++ oe_content.services.yml
@@ -4,9 +4,3 @@ services:
   oe_content.op_skos_setup:
     class: Drupal\oe_content\PublicationsOfficeSkosGraphSetup
     arguments: ['@config.factory']
-
-  oe_content.event_subscriber:
-    class: Drupal\oe_content\EventSubscriber\SubEntityAuthorSubscriber
-    parent: oe_content_sub_entity.event_subscriber.base
-    tags:
-      - { name: event_subscriber }
