+++ runner.yml.dist
@@ -42,7 +42,6 @@ drupal:
     - "./vendor/bin/drush en oe_content_consultation -y"
     - "./vendor/bin/drush en oe_content_person -y"
     - "./vendor/bin/drush en oe_content_organisation_person_reference -y"
-    - "./vendor/bin/drush en oe_content_sub_entity_extra_authors -y"
     - "./vendor/bin/drush cr"
     - "./vendor/bin/drush oe-multilingual:import-local-translations"
   settings:
